+++ src/components/Sidebar/NavLink.tsx
@@ -1,0 +1,29 @@
+import { Icon, Link, Text, LinkProps as ChakraLinkProps } from "@chakra-ui/react";
+import { ElementType } from "react";
+
+interface NavLinkProps extends ChakraLinkProps {
+  icon: ElementType
+  children: string
+}
+
+export function NavLink({ icon, children, ...rest }: NavLinkProps) {
+  return (
+    <Link
+      display="flex"
+      align="center"
+      {...rest}
+    >
+      <Icon
+        as={icon}
+        fontSize="20"
+      />
+
+      <Text
+        ml="4"
+        fontWeight="medium"
+      >
+        {children}
+      </Text>
+    </Link>
+  )
+}
